Fix #21 dockerwipe failure on no containers/images

+++ .bash_profile
@@ -156,17 +156,22 @@ dockerbuild() {
 }
 
 dockerwipe() {
+  # TODO: DRY up.
   local readonly containers=`docker ps -aq`
-  local readonly container_count=`$containers | wc -l`
-  echo -e "$COL_BLUE Removing Docker containers ($container_count)... $COL_RESET"
-  docker rm -v $containers
+  local readonly container_count=`docker ps -aq | wc -l | sed 's/[[:space:]]//g'`
+  if [ -n "$containers" ]; then
+    echo -e "$COL_BLUE Removing Docker containers ($container_count)... $COL_RESET"
+    docker rm -v $containers
+  fi
 
   local readonly images=`docker images --filter dangling=true -q`
-  local readonly image_count=`echo $images | wc -l`
-  echo -e "$COL_BLUE Removing dangling Docker images ($image_count)... $COL_RESET"
-  docker rmi $images
+  local readonly image_count=`docker images --filter dangling=true -q | wc -l | sed 's/[[:space:]]//g'`
+  if [ -n "$images" ]; then
+    echo -e "$COL_BLUE Removing dangling Docker images ($image_count)... $COL_RESET"
+    docker rmi $images
+  fi
 
-  echo -e "$COL_GREEN Docker wipe finished. $COL_RESET"
+  echo -e "$COL_GREEN Docker wipe finished. Removed $container_count container(s), $image_count image(s). $COL_RESET"
 }
 
 dockersize() {
